Simplify container start check

+++ docker/test/integration/minifi/core/DockerTestCluster.py
@@ -11,14 +11,8 @@ from .utils import retry_check
 
 class DockerTestCluster(SingleNodeDockerCluster):
     def __init__(self):
-        self.segfault = False
-        self.minifi_version = os.environ['MINIFI_VERSION']
-        self.minifi_root = '/opt/minifi/nifi-minifi-cpp-' + self.minifi_version
-        self.nifi_version = '1.7.0'
-        self.nifi_root = '/opt/nifi/nifi-' + self.nifi_version
-        self.kafka_broker_root = '/opt/kafka'
-
         super(DockerTestCluster, self).__init__()
+        self.segfault = False
 
     def deploy_flow(self):
         super(DockerTestCluster, self).deploy_flow()
@@ -32,9 +26,9 @@ class DockerTestCluster(SingleNodeDockerCluster):
             encoding = "utf8"
         return encoding
 
-    def get_app_log(self, container_id):
+    def get_app_log(self, container_name):
         try:
-            container = self.client.containers.get(container_id)
+            container = self.client.containers.get(container_name)
         except Exception:
             return 'not started', None
 
@@ -42,36 +36,37 @@ class DockerTestCluster(SingleNodeDockerCluster):
             logging.warn('Container segfaulted: %s', container.name)
             self.segfault = True
 
+        log_file_path = self.containers[container_name].get_log_file_path()
+        if not log_file_path:
+            return container.status, container.logs()
+
         try:
-            apps = [("MiNiFi", self.minifi_root + '/logs/minifi-app.log'), ("NiFi", self.nifi_root + '/logs/nifi-app.log')]
             if container.status == 'running':
-                for app in apps:
-                    app_log_status, app_log = container.exec_run('/bin/sh -c \'cat ' + app[1] + '\'')
-                    if app_log_status == 0:
-                        return container.status, app_log
+                app_log_status, app_log = container.exec_run('/bin/sh -c \'cat ' + log_file_path + '\'')
+                if app_log_status == 0:
+                    return container.status, app_log
             elif container.status == 'exited':
-                for app in apps:
-                    log_file_name = container_id + ".log"
-                    code = subprocess.run(["docker", "cp", container_id + ":" + app[1], log_file_name]).returncode
-                    if code == 0:
-                        output = open(log_file_name, 'rb').read()
-                        os.remove(log_file_name)
-                        return container.status, output
+                log_file_name = container_name + ".log"
+                code = subprocess.run(["docker", "cp", container_name + ":" + log_file_path, log_file_name]).returncode
+                if code == 0:
+                    output = open(log_file_name, 'rb').read()
+                    os.remove(log_file_name)
+                    return container.status, output
         except Exception:
             return container.status, None
 
-        return container.status, container.logs()
+        return container.status, None
 
-    def wait_for_app_logs(self, log, timeout_seconds, count=1):
+    def wait_for_app_logs(self, container_name, timeout_seconds, count=1):
         wait_start_time = time.perf_counter()
+        log = self.containers[container_name].get_startup_finish_text()
         while (time.perf_counter() - wait_start_time) < timeout_seconds:
-            for container_name in self.containers:
-                logging.info('Waiting for app-logs `%s` in container `%s`', log, container_name)
-                status, logs = self.get_app_log(container_name)
-                if logs is not None and count <= logs.decode("utf-8").count(log):
-                    return True
-                elif status == 'exited':
-                    return False
+            logging.info('Waiting for app-logs `%s` in container `%s`', log, container_name)
+            status, logs = self.get_app_log(container_name)
+            if logs is not None and count <= logs.decode("utf-8").count(log):
+                return True
+            elif status == 'exited':
+                return False
             time.sleep(1)
         return False
